Avoid importing any from apolloClientInstanceImport
Using the latest version our build is failing because in the generated graphql the context type is imported, but it's not set and thus defaults to `any`.

```
import { getApolloClient, any } from '../apolloClient';
```

This doesn't work, because we don't export something called any. I assume your intention is to use TypeScript `any` here?

This is just a quick fix to filter `any` out. I wasn't sure if you wanted to use `contextTypeRequired` instead, so I figured to stay as close to the code as-is now.